refactor: 清理冗余代码

+++ app/src/main/java/yangfentuozi/runner/app/service/ExecOnBootService.kt
@@ -41,7 +41,7 @@ class ExecOnBootService : Service() {
                 Log.i(TAG, "Shizuku is ready")
                 Runner.refreshStatus()
                 Runner.tryBindService()
-                if (Runner.waitService((timeout * 3).toLong())) {
+                if (Runner.waitService(timeout * 3)) {
                     Log.i(TAG, "Service is ready")
                     exec()
                 } else {
